u64 datatype for vmc

+++ modules/iopcore/cdvdman/device.h
@@ -7,4 +7,4 @@ void DeviceLock(void);    // Prevents further accesses to the device.
 void DeviceUnmount(void); // Called when OPL is shutting down.
 void DeviceStop(void);    // Called before the PS2 is to be shut down.
 
-int DeviceReadSectors(u32 lsn, void *buffer, unsigned int sectors);
+int DeviceReadSectors(u64 lsn, void *buffer, unsigned int sectors);
